backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
redis>=5.0.0
celery>=5.3.0
websockets>=12.0
pyyaml>=6.0
jinja2>=3.1.2
weasyprint>=60.0
openpyxl>=3.1.0
croniter>=2.0.0
paramiko>=3.0.0
